fix: revert test-harness.ts deduplication — depends on version-specific mock-api
test-harness.ts imports createMockAPI from ./mock-api which differs between
v0.6 and v0.7. The shared copy had a broken import since no mock-api.ts
existed in the shared directory. Reverted test-harness.ts back to version-
specific copies while keeping mock-context.ts and mock-agents.ts shared
(they have no version-specific imports).
