+++ sdk/shared/plugin-testing/package.json
@@ -5,8 +5,7 @@
   "description": "Shared testing utilities for Clubhouse plugin SDK versions",
   "exports": {
     "./mock-context": "./src/mock-context.ts",
-    "./mock-agents": "./src/mock-agents.ts",
-    "./test-harness": "./src/test-harness.ts"
+    "./mock-agents": "./src/mock-agents.ts"
   },
   "peerDependencies": {
     "@clubhouse/plugin-types": ">=0.6.0"
